# Add benchmark and improve tokenizer performance