--- a/sqltoken/tokenizer.go
+++ b/sqltoken/tokenizer.go
@@ -99,20 +99,44 @@ func ComparePos(x, y Pos) int {
 }
 
 type Tokenizer struct {
-	Dialect dialect.Dialect
-	Scanner *scanner.Scanner
-	Line    int
-	Col     int
+	Dialect      dialect.Dialect
+	Scanner      *scanner.Scanner
+	Line         int
+	Col          int
+	parseComment bool
 }
 
 func NewTokenizer(src io.Reader, dialect dialect.Dialect) *Tokenizer {
 	var scan scanner.Scanner
 	return &Tokenizer{
-		Dialect: dialect,
-		Scanner: scan.Init(src),
-		Line:    1,
-		Col:     1,
+		Dialect:      dialect,
+		Scanner:      scan.Init(src),
+		Line:         1,
+		Col:          1,
+		parseComment: true,
 	}
+}
+
+type TokenizerOption func(*Tokenizer)
+
+func Dialect(dialect dialect.Dialect) TokenizerOption {
+	return func(tokenizer *Tokenizer) {
+		tokenizer.Dialect = dialect
+	}
+}
+
+func DisableParseComment() TokenizerOption {
+	return func(tokenizer *Tokenizer) {
+		tokenizer.parseComment = false
+	}
+}
+
+func NewTokenizerWithOptions(src io.Reader, options ...TokenizerOption) *Tokenizer {
+	tokenizer := NewTokenizer(src, &dialect.GenericSQLDialect{})
+	for _, o := range options {
+		o(tokenizer)
+	}
+	return tokenizer
 }
 
 func (t *Tokenizer) Tokenize() ([]*Token, error) {
@@ -125,6 +149,10 @@ func (t *Tokenizer) Tokenize() ([]*Token, error) {
 		}
 		if err != nil {
 			return nil, err
+		}
+
+		if t == nil {
+			continue
 		}
 		tokenset = append(tokenset, t)
 	}
@@ -140,6 +168,10 @@ func (t *Tokenizer) NextToken() (*Token, error) {
 	}
 	if err != nil {
 		return &Token{Kind: ILLEGAL, Value: "", From: pos, To: t.Pos()}, errors.Errorf("tokenize failed: %w", err)
+	}
+
+	if !t.parseComment && (tok == Whitespace || tok == Comment) {
+		return nil, nil
 	}
 
 	return &Token{Kind: tok, Value: str, From: pos, To: t.Pos()}, nil