--- a/sqltoken/tokenizer.go
+++ b/sqltoken/tokenizer.go
@@ -437,13 +437,15 @@ func (t *Tokenizer) tokenizeWord(f rune) string {
 			break
 		}
 	}
+
 	str := builder.String()
 	t.Col += len(str)
 	return str
 }
 
 func (t *Tokenizer) tokenizeSingleQuotedString() (string, error) {
-	var str []rune
+	// var str []rune
+	var builder strings.Builder
 	t.Scanner.Next()
 
 	for {
@@ -451,7 +453,8 @@ func (t *Tokenizer) tokenizeSingleQuotedString() (string, error) {
 		if n == '\'' {
 			t.Scanner.Next()
 			if t.Scanner.Peek() == '\'' {
-				str = append(str, '\'')
+				// str = append(str, '\'')
+				builder.WriteRune('\'')
 				t.Scanner.Next()
 			} else {
 				break
@@ -459,15 +462,17 @@ func (t *Tokenizer) tokenizeSingleQuotedString() (string, error) {
 			continue
 		}
 		if n == scanner.EOF {
-			return "", errors.Errorf("unclosed single quoted string: %s at %+v", string(str), t.Pos())
+			return "", errors.Errorf("unclosed single quoted string: %s at %+v", builder.String(), t.Pos())
 		}
 
 		t.Scanner.Next()
-		str = append(str, n)
+		builder.WriteRune(n)
+		// str = append(str, n)
 	}
+	str := builder.String()
 	t.Col += 2 + len(str)
 
-	return string(str), nil
+	return str, nil
 }
 
 func (t *Tokenizer) tokenizeMultilineComment() (string, error) {