goos: darwin
goarch: amd64
pkg: github.com/akito0107/xsqlparser/sqltoken

```name                                      old time/op    new time/op    delta
Tokenizer_Tokenize/select-12                6.62µs ± 1%    5.21µs ± 0%  -21.33%  (p=0.000 n=9+8)
Tokenizer_Tokenize/complex_select-12        12.2µs ± 0%    10.1µs ± 0%  -16.69%  (p=0.000 n=10+8)
Tokenizer_Tokenize/insert-12                5.49µs ± 0%    5.67µs ±18%     ~     (p=0.497 n=9+10)
Tokenizer_Tokenize/multi_line_comment-12    16.5µs ± 1%    15.4µs ± 6%   -7.06%  (p=0.000 n=10+10)

name                                      old alloc/op   new alloc/op   delta
Tokenizer_Tokenize/select-12                6.22kB ± 0%    5.43kB ± 0%  -12.72%  (p=0.000 n=10+10)
Tokenizer_Tokenize/complex_select-12        11.8kB ± 0%    10.4kB ± 0%  -12.31%  (p=0.000 n=10+10)
Tokenizer_Tokenize/insert-12                5.88kB ± 0%    5.66kB ± 0%   -3.67%  (p=0.000 n=10+10)
Tokenizer_Tokenize/multi_line_comment-12    13.6kB ± 0%    12.6kB ± 0%   -7.53%  (p=0.000 n=10+10)

name                                      old allocs/op  new allocs/op  delta
Tokenizer_Tokenize/select-12                   112 ± 0%        79 ± 0%  -29.46%  (p=0.000 n=10+10)
Tokenizer_Tokenize/complex_select-12           195 ± 0%       150 ± 0%  -23.08%  (p=0.000 n=10+10)
Tokenizer_Tokenize/insert-12                   109 ± 0%        95 ± 0%  -12.84%  (p=0.000 n=10+10)
Tokenizer_Tokenize/multi_line_comment-12       302 ± 0%       246 ± 0%  -18.54%  (p=0.000 n=10+10)```